Move polling parameters from SnapPuller to ReplicationHandler, other refactoring
